feat(docs): keep hashed asset chunks out of the search index
Add an `X-Robots-Tag: noindex` rule for `/assets/*` to the Cloudflare
`_headers` file. Googlebot still fetches the chunks to render a page,
so a `robots.txt` Disallow would block rendering; the header alone
drops them from the results.

Move the `_headers` table out of v1-redirects.ts into its own
src/lib/headers.ts, with its tests, so the redirects module owns
redirects only.
