+++ apps/docs/src/lib/v1-redirects.ts
@@ -1,7 +1,7 @@
-// Redirect and header tables for the Cloudflare asset layer, expanded from the
+// Redirect table for the Cloudflare asset layer, expanded from the
 // route inventory in lib/v1-routes.ts. `vite.config.ts` emits the rendered
 // files into the client build so the asset layer answers them without a Worker
-// invocation.
+// invocation. The `_headers` file is lib/headers.ts.
 //
 // Query contract carried to every landing page (read by the redirect notice):
 //   lang=zh-CN — the request came from a /zh-CN path; v2 has no Chinese pages yet
@@ -25,12 +25,6 @@ export interface RedirectRule {
   destination: string;
   /** 308 for a permanent move, 307 while the target is still provisional. */
   status: 307 | 308;
-}
-
-/** One `_headers` block: a path pattern and the headers it adds. */
-export interface HeaderRule {
-  source: string;
-  headers: Record<string, string>;
 }
 
 const q = (src: string, lang?: string) =>
@@ -92,48 +86,10 @@ export function buildV1Redirects(): RedirectRule[] {
   return redirects;
 }
 
-export function buildHeaderRules(): HeaderRule[] {
-  return [
-    /**
-     * giscus loads the custom comment themes (public/giscus/*.css) into its
-     * cross-origin iframe via `<link crossorigin="anonymous">`, so they must
-     * send an Access-Control-Allow-Origin header or the browser blocks the
-     * stylesheet.
-     */
-    {
-      source: "/giscus/*",
-      headers: { "Access-Control-Allow-Origin": "https://giscus.app" },
-    },
-    // The asset layer types a prerendered `.xml` file as generic XML; feed
-    // readers expect the RSS media type the Next.js site served.
-    {
-      source: "/changelog/rss.xml",
-      headers: { "Content-Type": "application/rss+xml; charset=utf-8" },
-    },
-    // Every archive URL is pinned to the commit it was built from, so its
-    // bytes never change.
-    {
-      source: "/.well-known/agent-skills/*/archive.zip",
-      headers: { "Cache-Control": "public, max-age=31536000, immutable" },
-    },
-  ];
-}
-
 /** @see https://developers.cloudflare.com/workers/static-assets/redirects/ */
 export function renderRedirectsFile(rules = buildV1Redirects()): string {
   const lines = rules.map(
     ({ source, destination, status }) => `${source} ${destination} ${status}`,
   );
   return `${lines.join("\n")}\n`;
-}
-
-/** @see https://developers.cloudflare.com/workers/static-assets/headers/ */
-export function renderHeadersFile(rules = buildHeaderRules()): string {
-  const blocks = rules.map(({ source, headers }) =>
-    [
-      source,
-      ...Object.entries(headers).map(([name, value]) => `  ${name}: ${value}`),
-    ].join("\n"),
-  );
-  return `${blocks.join("\n\n")}\n`;
 }
