+++ apps/docs/vite.config.ts
@@ -11,13 +11,11 @@ import { defineConfig } from "vite";
 import type { Plugin } from "vite";
 
 import { agentSkillAssets } from "./src/lib/agent-skills.js";
+import { renderHeadersFile } from "./src/lib/headers.js";
 import { createOgCardRenderer } from "./src/lib/og-card.js";
 import { ogCards } from "./src/lib/og-cards.js";
 import { prerenderPages } from "./src/lib/prerender-pages.js";
-import {
-  renderHeadersFile,
-  renderRedirectsFile,
-} from "./src/lib/v1-redirects.js";
+import { renderRedirectsFile } from "./src/lib/v1-redirects.js";
 
 const packageRoot = import.meta.dirname;
 // Keep Miniflare's local Worker registry with the package's other ignored
@@ -70,6 +68,7 @@ function resolvedDocsLine(): Cloudflare.Env["DOCS_LINE"] {
  * Emits the Cloudflare asset-layer rule files into the client build, so legacy
  * permalinks and the giscus CORS header resolve without a Worker invocation.
  * @see src/lib/v1-redirects.ts
+ * @see src/lib/headers.ts
  */
 function cloudflareAssetRules(): Plugin {
   return {
